Further refactor code into separate crates
