+++ bins/server/Cargo.toml
@@ -12,20 +12,19 @@ rust-version = "1.68.0"
 
 [features]
 default = ["memcache"]
-memcache = ["dep:memcache", "dep:r2d2-memcache", "dep:r2d2"]
+memcache = ["dep:backend-memcache", "app-config/memcache"]
 
 [dependencies]
 anyhow = "1.0.75"
-async-tempfile = "0.5.0"
+app-config = { version = "0.1.0", path = "../../crates/app-config" }
 axum = { version = "0.6.20", features = ["http2", "headers", "macros", "json"] }
 backbone = { path = "../../crates/backbone" }
 backbone-traits = { path = "../../crates/backbone-traits" }
+backend-memcache = { version = "0.1.0", path = "../../crates/backend-memcache", optional = true }
 backend-traits = { version = "0.1.0", path = "../../crates/backend-traits" }
 base64 = "0.21.5"
-bytes = "1.5.0"
 chrono = "0.4.31"
 clap = { version = "4.4.11", features = ["env"] }
-config = "0.13.4"
 crossbeam = "0.8.2"
 ctrlc = { version = "3.4.1", features = ["termination"] }
 directories = "5.0.1"
@@ -36,19 +35,11 @@ futures = "0.3.29"
 headers-content-md5 = "0.1.1"
 hex = "0.4.3"
 hyper = { version = "0.14.27", features = ["http1", "http2", "server", "h2"] }
-lazy_static = "1.4.0"
-map-ok = "0.1.0"
-memcache = { version = "0.17.1", optional = true }
 metrics = { version = "0.1.0", path = "../../crates/metrics" }
 mime-db = "1.7.0"
 percent-encoding = "2.3.1"
 pin-project = "1.1.3"
 problemdetails = { version = "0.2.1", features = ["axum"] }
-prometheus-client = "0.22.0"
-prost = "0.12.3"
-prost-derive = "0.12.3"
-r2d2 = { version = "0.8.10", optional = true }
-r2d2-memcache = { version = "0.6.0", optional = true }
 rendezvous = { version = "0.2.2", features = ["tokio", "log"] }
 serde = { version = "1.0.193", features = ["derive"] }
 shared-files = "0.1.0"
@@ -62,12 +53,8 @@ tracing = "0.1.40"
 tracing-subscriber = { version = "0.3.18", features = ["env-filter", "parking_lot", "tracing-log", "json"] }
 uuid = { version = "1.6.1", features = ["v1", "rng", "serde"] }
 
-
 [dev-dependencies]
 serde_yaml = "0.9.27"
-
-[build-dependencies]
-prost-build = "0.12.3"
 
 [package.metadata.docs.rs]
 all-features = true
