--- bins/server/src/backends/memcache.rs
@@ -1,7 +1,0 @@
-mod backend;
-mod config;
-mod connection_string;
-
-pub use backend::{MemcacheBackend, MemcacheBackendConstructionError};
-pub use config::MemcacheBackendConfig;
-pub use connection_string::MemcacheConnectionString;
