--- bins/server/src/backends/mod.rs
@@ -1,6 +1,0 @@
-pub use crate::backends::registry::{BackendRegistry, TryCreateFromConfig};
-
-#[cfg_attr(docsrs, doc(cfg(feature = "memcache")))]
-#[cfg(feature = "memcache")]
-pub mod memcache;
-mod registry;
