+++ crates/app-config/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "app-config"
+version = "0.1.0"
+edition = "2021"
+
+[features]
+memcache = []
+
+[dependencies]
+clap = "4.4.11"
+config = "0.13.4"
+serde = "1.0.193"
+tracing = "0.1.40"
+anyhow = "1.0.75"
+matches = "0.1.10"
+thiserror = "1.0.50"
+url = "2.5.0"
+
+[dev-dependencies]
+serde_yaml = "0.9.27"
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
