+++ crates/backbone/Cargo.toml
@@ -19,3 +19,7 @@ shortguid = "0.6.0"
 thiserror = "1.0.50"
 tokio = { version = "1.35.0", features = ["io-std", "time"] }
 tracing = "0.1.40"
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
