+++ crates/backbone/src/lib.rs
@@ -1,3 +1,7 @@
+// only enables the `doc_cfg` feature when
+// the `docsrs` configuration attribute is defined
+#![cfg_attr(docsrs, feature(doc_cfg))]
+
 mod backbone;
 mod file_accessor;
 mod file_reader;
