+++ crates/backend-traits/Cargo.toml
@@ -3,12 +3,15 @@ name = "backend-traits"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
+app-config = { version = "0.1.0", path = "../app-config" }
 async-trait = "0.1.74"
 backbone-traits = { version = "0.1.0", path = "../backbone-traits" }
 file_distribution = { version = "0.1.0", path = "../file_distribution" }
 shortguid = "0.6.0"
 thiserror = "1.0.50"
 tokio = { version = "1.35.0", default-features = false, features = ["rt", "sync"] }
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
