+++ crates/backend-traits/src/backend.rs
@@ -1,0 +1,32 @@
+use async_trait::async_trait;
+use backbone_traits::{FileAccessor, FileAccessorError};
+use file_distribution::WriteSummary;
+use shortguid::ShortGuid;
+use std::error::Error;
+use std::sync::Arc;
+
+#[async_trait]
+pub trait Backend: Send + Sync {
+    /// Gets the tag of the backend.
+    fn tag(&self) -> &str;
+
+    /// Handles a file that is ready for distribution.
+    async fn distribute_file(
+        &self,
+        id: ShortGuid,
+        summary: Arc<WriteSummary>,
+        file_accessor: Arc<dyn FileAccessor>,
+    ) -> Result<(), DistributionError>;
+}
+
+#[derive(Debug, thiserror::Error)]
+pub enum DistributionError {
+    #[error(transparent)]
+    BackendSpecific(Box<dyn Error>),
+    #[error(transparent)]
+    FileAccessor(#[from] FileAccessorError),
+    #[error(transparent)]
+    Io(#[from] std::io::Error),
+    #[error(transparent)]
+    Join(#[from] tokio::task::JoinError),
+}
