+++ crates/backend-traits/src/backend_info.rs
@@ -1,0 +1,9 @@
+pub trait BackendInfo {
+    /// Gets a short name of the backend.
+    fn backend_name() -> &'static str;
+
+    /// Gets an informational string about the backend.
+    fn backend_version() -> &'static str {
+        ""
+    }
+}
