+++ crates/file_distribution/Cargo.toml
@@ -4,7 +4,17 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+bytes = "1.5.0"
 md5 = "0.7.0"
+prost = "0.12.3"
+prost-derive = "0.12.3"
 sha2 = "0.10.8"
 shortguid = "0.6.0"
 tokio = { version = "1.35.0", default-features = false, features = ["time"] }
+
+[build-dependencies]
+prost-build = "0.12.3"
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
