+++ crates/metrics/src/lib.rs
@@ -1,5 +1,9 @@
 //! Contains metrics related code.
 
+// only enables the `doc_cfg` feature when
+// the `docsrs` configuration attribute is defined
+#![cfg_attr(docsrs, feature(doc_cfg))]
+
 pub mod http;
 pub mod transfer;
